Refactor area-related functions
